For the hal_flash API's make the change to uintptr_t to support 64-bit explicit using `ARCH_64BIT` or known 64-bit architectures.

+++ hal/nxp_t2080.c
@@ -532,7 +532,7 @@ void hal_init(void)
 #endif
 }
 
-int hal_flash_write(uintptr_t address, const uint8_t *data, int len)
+int hal_flash_write(uint32_t address, const uint8_t *data, int len)
 {
     (void)address;
     (void)data;
@@ -541,7 +541,7 @@ int hal_flash_write(uintptr_t address, const uint8_t *data, int len)
     return 0;
 }
 
-int hal_flash_erase(uintptr_t address, int len)
+int hal_flash_erase(uint32_t address, int len)
 {
     (void)address;
     (void)len;
